feat(spending goals): show actual expenses beside target
In the sumary bar, the target spending, as set by the user, is updated
each time a transaction of type `Income` is added.
Beside the target spending, in parentheses, the sum of actual spending
for each goal is shown, (calculated using the category and subcategory
of each new transaction of type `Expense` when is added).