Update dependencies and version for budget_tracker_tui
Bumped the version of the `budget_tracker_tui` package to 1.2.0 and updated several dependencies, including `csv`, `serde`, `chrono`, and `serde_json` to their latest versions for improved functionality and performance. Added new dependencies such as `ureq` and `semver` to enhance the application's capabilities. Updated the `winresource` dependency to the latest version for better compatibility. Adjusted the package metadata version and copyright information accordingly.